TestRig v1.16
Functional Code, with solar charger code/hardware operating correctly unloaded.
Writing to files spaghetti removed to allow for multiple drives to be added in the future.
Latch results file created to easily debug an incorrect latch test.
Sample voltages and Input voltages monitored, added to the results file.
Results now stored into a separate directory containing all result files

+++ Core/Src/SerialNumber.c
@@ -148,7 +148,7 @@ bool ContinueWithCurrentSerial() {
 	LCD_displayString((uns_ch*) &lcdBuffer, strlen((char*) lcdBuffer));
 
 	LCD_ClearLine(3);
-	sprintf((char*) &lcdBuffer[0], "      %lu", (unsigned long) BoardConnected.SerialNumber);
+	sprintf((char*) &lcdBuffer[0], "      %lu", BoardConnected.SerialNumber);
 	printT((uns_ch*) &lcdBuffer[0]);
 	LCD_setCursor(3, 0);
 	LCD_displayString((uns_ch*) &lcdBuffer[0], strlen((char*) lcdBuffer));
